backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --format requirements-txt --no-hashes -o requirements.txt
annotated-types==0.7.0
    # via pydantic
asgiref==3.11.1
    # via
    #   django
    #   django-cors-headers
    #   django-ninja-extra
cffi==2.0.0 ; platform_python_implementation != 'PyPy'
    # via cryptography
contextlib2==21.6.0
    # via django-ninja-extra
cryptography==46.0.6
    # via pyjwt
dj-database-url==3.1.2
    # via workmy
django==6.0.3
    # via
    #   dj-database-url
    #   django-cors-headers
    #   django-ninja
    #   django-ninja-extra
    #   django-ninja-jwt
    #   workmy
django-cors-headers==4.9.0
    # via workmy
django-ninja==1.6.2
    # via
    #   django-ninja-extra
    #   workmy
django-ninja-extra==0.31.3
    # via django-ninja-jwt
django-ninja-jwt==5.4.4
    # via workmy
django-ratelimit==4.1.0
    # via workmy
gunicorn==25.3.0
    # via workmy
injector==0.24.0
    # via django-ninja-extra
packaging==26.0
    # via gunicorn
pillow==12.1.1
    # via workmy
psycopg2-binary==2.9.11
    # via workmy
pycparser==3.0 ; implementation_name != 'PyPy' and platform_python_implementation != 'PyPy'
    # via cffi
pydantic==2.12.5
    # via
    #   django-ninja
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.13.1
    # via django-ninja-jwt
pyjwt==2.12.1
    # via django-ninja-jwt
python-dotenv==1.2.2
    # via
    #   pydantic-settings
    #   workmy
sqlparse==0.5.5
    # via django
typing-extensions==4.15.0
    # via
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.3 ; sys_platform == 'win32'
    # via django
whitenoise==6.12.0
    # via workmy
